@@ -15,26 +15,28 @@
   "author": "Masashi Shibata",
   "license": "MIT",
   "dependencies": {
-    "@material-ui/core": "4.11.3",
-    "@material-ui/icons": "4.11.2",
-    "axios": "0.21.1",
-    "notistack": "1.0.1",
-    "plotly.js-dist": "1.57.1",
-    "react": "16.14.0",
-    "react-dom": "16.14.0",
-    "react-router-dom": "5.2.0",
-    "recoil": "0.1.2"
+    "@emotion/react": "^11.10.8",
+    "@emotion/styled": "^11.10.8",
+    "@mui/icons-material": "^5.11.6",
+    "@mui/lab": "^5.0.0-alpha.128",
+    "@mui/material": "^5.12.1",
+    "axios": "1.2.1",
+    "notistack": "3.0.1",
+    "plotly.js-dist": "2.22.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router-dom": "6.11.0",
+    "recoil": "0.7.7"
   },
   "devDependencies": {
-    "@babel/core": "7.12.10",
-    "@types/plotly.js": "1.54.7",
-    "@types/react": "16.9.53",
-    "@types/react-dom": "16.9.8",
-    "@types/react-router-dom": "5.1.6",
-    "prettier": "2.1.2",
-    "ts-loader": "8.4.0",
-    "typescript": "4.0.3",
-    "webpack": "5.20.1",
-    "webpack-cli": "4.1.0"
+    "@babel/core": "7.14.5",
+    "@types/plotly.js": "^2.12.11",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "prettier": "2.5.1",
+    "ts-loader": "9.2.7",
+    "typescript": "4.6.2",
+    "webpack": "5.76.0",
+    "webpack-cli": "4.9.2"
   }
 }
